Video Support in V1.0: part 1/3 : support matroskavideo : replace long by size_t and fix rebase issue

+++ include/exiv2/matroskavideo.hpp
@@ -110,7 +110,7 @@ class EXIV2API MatroskaVideo : public Image {
     @param buf Pointer to the memory area with the tag information.
     @param size Size of \em buf.
    */
-  void contentManagement(const MatroskaTags* mt, const byte* buf, long size);
+  void contentManagement(const MatroskaTags* mt, const byte* buf, size_t size);
   /*!
     @brief Calculates Aspect Ratio of a video, and stores it in the
         respective XMP container.
